microservices/bListener/requirements.txt
flask==1.1.1
python-dotenv==0.10.3
google-cloud-logging